Add default condition in formatResponse.
Putting 'application/json' as the first condition ensures that
when a */* Accept header comes in, JSON is returned. I want to
return JSON by default, so this will work if a browser string
comes in as the Accept header. But, if someone hits the route with
an Accept header like 'text/html', no response was returned, so
I included a 'default' condition that returns JSON.