Store middleware sub-stack for POST /email route in an array.
Once there are a few different middleware functions running in a
route-handler, it makes sense for readability to break them out
into an array, and pass that named array in as the sole param
after the mount point. This looks cleaner to me, but is not necessary
for all routes.